# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PaginationInterfaces": {
				"properties": {
					"meta": {
						"properties": {
							"total_pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"per_page": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total_pages",
							"total",
							"per_page",
							"page"
						],
						"type": "object"
					},
					"data": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"meta",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChoicesInterfaces": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"answer_id": {
						"type": "string"
					},
					"option_id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ChoicesCreateInterfaces": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"answer_id": {
						"type": "string"
					},
					"option_id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"answer_id",
					"option_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChoicesUpdateInterfaces": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					},
					"answer_id": {
						"type": "string"
					},
					"option_id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ChoicesDeleteInterfaces": {
				"properties": {
					"uuid": {
						"type": "string"
					}
				},
				"required": [
					"uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChoicesUploadDeleteInterfaces": {
				"properties": {
					"links": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"OptionsUploadDeleteInterfaces": {
				"properties": {
					"links": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"QuestionsUploadDeleteInterfaces": {
				"properties": {
					"links": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RoomsUploadDeleteInterfaces": {
				"properties": {
					"links": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"links"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-api-starter-ts",
		"version": "1.2.0",
		"description": " A basic starter for an express.js API with Typescript",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "CJ R. ",
			"email": "cj@null.computer",
			"url": "https://w3cj.now.sh"
		}
	},
	"paths": {
...
		"/api/v1/choices": {
			"get": {
				"operationId": "GetData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationInterfaces"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user.\nSupply the unique user ID from either and receive corresponding user details.",
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "uuid_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_exact",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_except",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_like",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_ilike",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_startswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_istartswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_endswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "uuid_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "name_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "description_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "is_active_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "boolean"
							}
						}
					},
					{
						"in": "query",
						"name": "answer_id_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "option_id_iendswith",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_gt",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_gt",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_gte",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_gte",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_lt",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_lt",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "created_at_lte",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "updated_at_lte",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string",
								"format": "date-time"
							}
						}
					},
					{
						"in": "query",
						"name": "order_by_asc",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "order_by_desc",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "CreateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChoicesInterfaces"
								}
							}
						}
					}
				},
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChoicesCreateInterfaces"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChoicesInterfaces"
								}
							}
						}
					}
				},
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChoicesUpdateInterfaces"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChoicesInterfaces"
								}
							}
						}
					}
				},
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChoicesDeleteInterfaces"
							}
						}
					}
				}
			}
		},
		"/api/v1/choices/upload": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Choices"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChoicesUploadDeleteInterfaces"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/options/upload": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Options"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Options"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OptionsUploadDeleteInterfaces"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/questions/upload": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Question"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuestionsUploadDeleteInterfaces"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/rooms/upload": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoomsUploadDeleteInterfaces"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}